separate mongo functions by scope
