+++ packages/database-mongo/package.json
@@ -10,7 +10,7 @@
     "precompile": "yarn clean",
     "compile": "tsc",
     "prepublishOnly": "yarn compile",
-    "testonly": "jest --runInBand --forceExit",
+    "testonly": "jest --runInBand",
     "test:watch": "jest --watch",
     "coverage": "yarn testonly --coverage"
   },
@@ -20,7 +20,8 @@
   ],
   "jest": {
     "testEnvironment": "node",
-    "preset": "ts-jest"
+    "preset": "ts-jest",
+    "testPathIgnorePatterns": ["__tests__/test-utils.ts"]
   },
   "repository": {
     "type": "git",
@@ -31,14 +32,14 @@
   "devDependencies": {
     "@accounts/database-tests": "^0.21.1",
     "@types/jest": "24.0.25",
-    "@types/lodash": "4.14.149",
+    "@types/lodash.merge": "4.6.6",
     "@types/mongodb": "3.3.14",
     "@types/node": "12.12.14",
     "jest": "24.9.0"
   },
   "dependencies": {
     "@accounts/types": "^0.21.1",
-    "lodash": "^4.17.15",
+    "lodash.merge": "^4.6.2",
     "mongodb": "^3.4.1",
     "tslib": "1.10.0"
   }
