+++ packages/database-mongo/src/options.ts
@@ -1,0 +1,12 @@
+export const defaultOptions = {
+  collectionName: 'users',
+  sessionCollectionName: 'sessions',
+  timestamps: {
+    createdAt: 'createdAt',
+    updatedAt: 'updatedAt',
+  },
+  convertUserIdToMongoObjectId: true,
+  convertSessionIdToMongoObjectId: true,
+  caseSensitiveUserName: true,
+  dateProvider: (date?: Date) => (date ? date.getTime() : Date.now()),
+};
