+++ packages/database-mongo/src/utils.ts
@@ -1,0 +1,8 @@
+import { ObjectID } from 'mongodb';
+
+export const toMongoID = (objectId: string | ObjectID) => {
+  if (typeof objectId === 'string') {
+    return new ObjectID(objectId);
+  }
+  return objectId;
+};
